MDEV-14425 WIP: Redo log format with arbitrary block size
The InnoDB redo log used to be formatted in blocks of 512 bytes.
The log blocks were encrypted and the checksum was calculated while
holding log_sys.mutex, creating a serious scalability bottleneck.

This work removes the redo log block structure altogether and
essentially turns every mini-transaction into a log block of its own.
This allows encryption and checksum calculations to be performed
in mtr_t::m_log, before acquiring log_sys.mutex. The mutex only
protects a memcpy() of the data to the shared log_sys.buf, as well
as the padding of the log, in case the to-be-written part of the log
would not end in a block boundary of the underlying storage.

We extend the MDEV-12353 redo log record format as follows.

(1) Empty mini-transactions or extra NUL bytes will not be allowed.
(2) The end-of-minitransaction marker (a NUL byte) will be replaced
with a 1-bit sequence number, which will be toggled each time when the
circular log file wraps back to the beginning.
(3) After the sequence bit, a CRC-32C checksum of all data
(excluding the sequence bit) will written.
(4) If the log is encrypted, 8 bytes will be written after
the checksum. This is part of the initialization vector (IV)
of encrypted log data.
(5) File names, page numbers, and checkpoint information will not be
encrypted. Only the payload bytes of page-level log will be encrypted.
The tablespace ID and page number will form part of the IV.
(6) For padding, arbitrary-length FILE_CHECKPOINT records may be written,
with all-zero payload, and with the normal end marker and checksum.

The 512-byte log checkpoint blocks will be moved from 0x200 and 0x600
will be replaced with 64-byte checkpoint blocks at 0x1000 and 0x2000.

The start of log records will move from 0x800 to 0x3000. This allows us
to use 4096-byte aligned blocks for all I/O.

FIXME: Fix and enable ./mtr --suite=mariabackup

FIXME: When reading or writing log blocks, use the file system
block size instead of using arbitrary byte offsets.

FIXME: Remove recv_sys_t::buf, and use log_sys.buf for parsing.

recv_sys_t::MTR_SIZE_MAX: Replaces RECV_SCAN_SIZE

recv_sys_t, log_sys_t: Removed many data members.

+++ storage/innobase/include/srv0srv.h
@@ -296,7 +296,7 @@ extern ulonglong	srv_log_file_size;
 extern ulong	srv_log_buffer_size;
 extern ulong	srv_flush_log_at_trx_commit;
 extern uint	srv_flush_log_at_timeout;
-extern ulong	srv_log_write_ahead_size;
+extern uint	srv_log_write_ahead_size;
 extern my_bool	srv_adaptive_flushing;
 extern my_bool	srv_flush_sync;
